feat: implemented documents resource, information group resource and information group's assemblers

+++ src/main/java/com/deltatech/diligencetech/platform/duediligence/domain/model/queries/GetInformationGroupByIdQuery.java
@@ -1,4 +1,4 @@
 package com.deltatech.diligencetech.platform.duediligence.domain.model.queries;
 
-public record GetInformationGroupByIdQuery(int id) {
+public record GetInformationGroupByIdQuery(Long id) {
 }
